fix: Strip spaces from secret data
The README shows how to create the cloudflare secret using literals,
which may persist in shell history. Another method is to write the
values to files (using `--from-file=`) and `shred` the files after the
secret is created. This currently doesn't work because a newline can be
appended to the end of the values. Spaces and newlines are not valid
characters for cloudflare accounts or tokens, so we can safely trim
them when reading the secret values. This change makes the secrets more
flexible and robust to user errors.