fix: Allow cloudflared ICMP proxy feature
When starting the cloudflared pod, the following is warning appears in
the logs:

```text
WRN The user running cloudflared process has a GID (group ID) that is not within ping_group_range. You might need to add that user to a group within that range, or instead update the range to encompass a group the user is already in by modifying /proc/sys/net/ipv4/ping_group_range. Otherwise cloudflared will not be able to ping this network error="Group ID 0 is not between ping group 1 to 0"
WRN ICMP proxy feature is disabled error="cannot create ICMPv4 proxy: Group ID 0 is not between ping group 1 to 0 nor ICMPv6 proxy: socket: permission denied"
```

Added the needed sysctl settings to the security context to enable this
feature and confirmed that the warning is removed.

https://github.com/cloudflare/cloudflared/issues/1109